+++ types/ui/vscode/index.d.ts
@@ -1,0 +1,6 @@
+/**
+ * Entry point for the VS Code plugin (extension)
+ * @param {vscode.ExtensionContext} context
+ */
+export function activate(context: vscode.ExtensionContext): void;
+export function deactivate(): void;